Central Files
An html file, with two css files which can be used to dictate the formatting of the document. The original CSS sheet is my attempt before watching the solution video and is referenced in the html document. The alternative has adjustments made to accommodate the small differences between my attempt and the example.